chore: move shared GSON instance into utils package
