+++ plugin-core/src/main/java/appland/installGuide/analyzer/languages/JavaScriptLanguageAnalyzer.java
@@ -1,6 +1,7 @@
 package appland.installGuide.analyzer.languages;
 
 import appland.installGuide.analyzer.*;
+import appland.utils.GsonUtils;
 import com.google.gson.JsonObject;
 import com.intellij.openapi.fileEditor.FileDocumentManager;
 import com.intellij.openapi.util.text.StringUtil;
