+++ plugin-core/src/main/java/appland/utils/GsonUtils.java
@@ -1,4 +1,4 @@
-package appland.installGuide.analyzer;
+package appland.utils;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
